Add the ccap-direct example code
Introduced the ccap-direct example code, showcasing how to use the CCAP driver
to directly render captured images to specified positions on the frame buffer.
This example forks a process to run GStreamer, which drives the CCAP driver
for image capture. Simultaneously, another process uses ioctl to control the
CCAP driver.
The example demonstrates not only direct rendering of captured images to the
frame buffer but also the functionality of dynamically changing the output
position.

Change-Id: If1d53d5ed20c14b23ba27702f268a2146d2e1f31